fix: Resolve CORS Errors by Using Same Domain for API Calls
We didn't run into this issue on our Linux machine, but
we have run into it with our MacBook (which is nice because
it means we got to prevent user headaches early). From
now on, all API calls made from the frontend must be
to the same domain to prevent CORS issues. (That is,
use `/todos`, not `http://localhost:5001/todos` on
the frontend.)

Also updated tests.